refactor circular dependency issues
